sc-babe: pub api to construct BabeVerifier
By default in Polkadot SDK, BabeVerifiers can only be constructed in the
same file the struct is declared.

This commit exposes a public `fn new` constructor method for the
BabeVerifier. This is useful for the OTF Hybrid Import Queue, which
constructs both an Aura and Babe verifier to use internally.