sc-babe: allow parent with non-zero block number to be `UnimportedGenesis`
By default, aux_data lookups hard code checking if the block number is
zero to return UnimportedGenesis, and otherwise will return an error if
parent does not contain a Babe predigest.

This hardcoded check of block zero does not work for live chains that
perform Aura to Babe runtime upgrades.

It is not possible within the `fn epoch_descriptor_for_child_of` call to
check if this is the first Babe block to be mined without significant
breaking changes. Luckily, it is safe to simply make the assumption that
if parent epoch changes are missing, this is the first Babe block.
Rationale for why this is safe is commented above the changed code.