sc-babe: remove bad hardcoding in `fn find_pre_digest`
`fn find_pre_digest` states in its doc-comment: "Pre-runtime digests are
mandatory, the function will return Err if none is found.".

However makes an exception to this rule when the parent block number is
zero, instead returning a "Default" value for the digest when the header
block is zero.

Because after an Aura to Babe runtime upgrade is possible for the parent
pre-runtime digest to be missing, I have modified the return value of
this fn to `Result<Option<PreDigest>, ...>`, returning `Ok(None)` if the
pre-runtime digest is missing and allowing the caller to decide what to
do in that case. Most of the time, the caller will choose to
.unwrap_or_default.